Refactor: extract generic story creation logic to Story class
