+++ home/.bin/hotfix.rb
@@ -1,16 +1,6 @@
-require 'shellwords'
-require 'yaml'
-require 'active_support'
-require_relative 'git_repository'
-require_relative 'pivotal_project'
+require_relative 'story'
 
-class Hotfix
-  def initialize(story_name:, project_id:, pivotal_token:)
-    @story_name = story_name
-    @project_id = project_id
-    @pivotal_token = pivotal_token
-  end
-
+class Hotfix < Story
   def save!
     validate_environment_and_arguments
     validate_git_repository
@@ -19,26 +9,6 @@ class Hotfix
   end
 
   private
-
-  attr_reader :story_name, :project_id, :pivotal_token
-
-  def validate_environment_and_arguments
-    if env_variable_undefined?(story_name)
-      raise ArgumentError, "please pass a pivotal story name"
-    end
-
-    if env_variable_undefined?(project_id)
-      raise ArgumentError, "You need to populate the PIVOTAL_PROJECT_ID environment variable"
-    end
-
-    if env_variable_undefined?(pivotal_token)
-      raise ArgumentError, "You need to populate the PIVOTAL_TOKEN environment variable"
-    end
-  end
-
-  def env_variable_undefined?(variable)
-    variable.nil? || variable.empty?
-  end
 
   def validate_git_repository
     GitRepository.new(Dir.pwd).validate_for_hotfix
@@ -53,19 +23,6 @@ class Hotfix
 
     escaped_hotfix_name = create_hotfix_name(story.id)
     `git flow hotfix start #{escaped_hotfix_name}`
-  end
-
-  def pivotal_project
-    PivotalProject.new(pivotal_token, project_id)
-  end
-
-  def derive_story_owners
-    hitch_config = YAML.load_file("#{ENV['HOME']}/.hitchrc")
-    if hitch_config[:current_pair].empty?
-      Array(ENV['USER'])
-    else
-      hitch_config[:current_pair]
-    end
   end
 
   def create_hotfix_name(story_id)
